// frontend/src/app/models/report/reportCoursePaymentData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ReportCoursePaymentData {
    courseCode?: string;
    first?: number;
    fullCourseEn?: string;
    fullCourseTh?: string;
    mode?: string;
    payAmount?: number;
    publicNameEn?: string;
    publicNameTh?: string;
    refundAmount?: number;
    registerAmount?: number;
    registerCount?: number;
    remainAmount?: number;
    rowNum?: number;
    size?: number;
}